kali_automation/requirements.txt
# Complete Kali Linux Automation Platform Requirements
# Core automation and orchestration
celery>=5.3.0
redis>=4.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Network and security tools
python-nmap>=0.7.1
scapy>=2.5.0
impacket>=0.10.0
pwntools>=4.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.2.0

# Machine learning and AI
scikit-learn>=1.3.0
tensorflow>=2.14.0
torch>=2.1.0
transformers>=4.35.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.6.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0
loguru>=0.7.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Additional security tools
yara-python>=4.3.0
pefile>=2023.2.7
capstone>=4.0.2
keystone-engine>=0.9.2
unicorn>=2.0.1
